refactor(web): SectionHeaderHOC 제거 및 SectionHeader로 통합
