+++ apps/web/src/app/(desktop-ready)/(home)/components/LiveHotDealSection.tsx
@@ -2,7 +2,7 @@ import { Suspense } from 'react';
 
 import ApiErrorBoundary from '@/components/ApiErrorBoundary';
 import { IllustStanding } from '@/components/common/icons';
-import SectionHeader from '@/components/SectionHeaderHOC';
+import SectionHeader from '@/components/SectionHeader';
 
 import LiveHotDealList from './LiveHotDealList';
