+++ apps/web/src/app/(desktop-ready)/recommend/layout.tsx
@@ -2,7 +2,7 @@ import { ReactNode } from 'react';
 
 import { checkDevice } from '@/app/actions/agent';
 import BasicLayout from '@/components/layout/BasicLayout';
-import SectionHeaderHOC from '@/components/SectionHeaderHOC';
+import SectionHeader from '@/components/SectionHeader';
 
 import RecommendPageHeader from './components/RecommendPageHeader';
 
@@ -15,7 +15,7 @@ export default async function Layout({ children }: { children: ReactNode }) {
         <BasicLayout header={<RecommendPageHeader />}>{children}</BasicLayout>
       ) : (
         <div className="mt-14 pt-8">
-          <SectionHeaderHOC title="지름알림 추천" />
+          <SectionHeader title="지름알림 추천" />
           <div className="mx-auto max-w-screen-layout-max">{children}</div>
         </div>
       )}
