+++ apps/web/src/components/SectionHeader.tsx
@@ -6,20 +6,17 @@ const SectionHeader = ({
   title,
   right,
   titleClassName,
-  isMobile = true,
 }: {
   title: ReactNode;
   right?: ReactNode;
   titleClassName?: string;
-  isMobile?: boolean;
 }) => {
   return (
     <div className="relative flex h-[56px] w-full items-center justify-between">
-      {!isMobile && <div className="hidden pc:block" />}
+      <div className="hidden pc:block" />
       <h2
         className={cn(
-          'text-lg font-bold text-gray-900',
-          !isMobile && 'pc:absolute pc:left-1/2 pc:-translate-x-1/2 pc:text-[28px]',
+          'text-lg font-bold text-gray-900 pc:absolute pc:left-1/2 pc:-translate-x-1/2 pc:text-[28px]',
           titleClassName,
         )}
       >
