--- apps/web/src/components/SectionHeaderHOC.tsx
@@ -1,6 +1,0 @@
-import withCheckDevice from './hoc/withCheckDevice';
-import SectionHeader from './SectionHeader';
-
-const SectionHeaderHOC = withCheckDevice(SectionHeader);
-
-export default SectionHeaderHOC;
